creation of UserRepository, corrections in Order, OrderRepository and OrderRepositoryTestSuite

+++ src/main/java/com/kodilla/ecommercee/repository/OrderRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface OrderRepository extends CrudRepository<Order, Integer> {
@@ -13,6 +12,6 @@ public interface OrderRepository extends CrudRepository<Order, Integer> {
     @Override
     List<Order> findAll();
 
-    Order findById(int integer);
+    Order findById(Long integer);
 
 }
